build(api): fix esm issues
Workaround NCC ESM issues by just not using ESM with NCC.
The API is still ESM, but tsup converts it to CJS before NCC consumes it, as a temporary workaround until there is something to replace NCC. Please save us bun 🙏
This seems to work fine, and as a free bonus, there is a huge build speedup because NCC isn't trying to run tsc.

Shoutout to ESM! Because forcing everyone to choose between ESM, where you need multiple workarounds a day on every package imaginable to get it to work, and CJS, where you opt yourself out of a large and growing number of NPM packages, is a fantastic idea. Interoperability is for fools!